Migrate GetNotificationDocumentMetadataRecord
