+++ src/domain/Repository.ts
@@ -10,6 +10,7 @@ import { ConsumeEventStreamRecord } from './ConsumeEventStreamRecord';
 import { CheckNotificationStatusRecord } from './CheckNotificationStatusRecord';
 import { ListEventStreamRecord } from './ListEventStreamRecord';
 import { GetNotificationDetailRecord } from './GetNotificationDetailRecord';
+import { GetNotificationDocumentMetadataRecord } from './GetNotificationDocumentMetadataRecord';
 
 export type Repository<A> = {
   insert: (input: A) => TE.TaskEither<Error, A>;
@@ -41,7 +42,8 @@ export type Record =
   | ConsumeEventStreamRecord
   | CreateEventStreamRecord
   | ListEventStreamRecord
-  | GetNotificationDetailRecord;
+  | GetNotificationDetailRecord
+  | GetNotificationDocumentMetadataRecord;
 
 export type RecordRepository = {
   insert: <A extends Record>(input: A) => TE.TaskEither<Error, A>;
