+++ src/useCases/SystemEnv.ts
@@ -1,4 +1,3 @@
-import { GetNotificationDocumentMetadataRecordRepository } from '../domain/GetNotificationDocumentMetadataRepository';
 import { GetPaymentNotificationMetadataRecordRepository } from '../domain/GetPaymentNotificationMetadataRecordRepository';
 import { DomainEnv } from '../domain/DomainEnv';
 import { LegalFactDownloadMetadataRecordRepository } from '../domain/LegalFactDownloadMetadataRecordRepository';
@@ -8,7 +7,6 @@ import { RecordRepository } from '../domain/Repository';
 export type SystemEnv = DomainEnv & {
   // repositories
   recordRepository: RecordRepository;
-  getNotificationDocumentMetadataRecordRepository: GetNotificationDocumentMetadataRecordRepository;
   getPaymentNotificationMetadataRecordRepository: GetPaymentNotificationMetadataRecordRepository;
   getLegalFactDownloadMetadataRecordRepository: LegalFactDownloadMetadataRecordRepository;
   getNotificationPriceRecordRepository: GetNotificationPriceRecordRepository;
